# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ecommerce API",
    "description": "API for managing products and categories",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "description": "Retrieve all products",
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "OK" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Products"],
        "description": "Create a new product",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "example": "Gaming Laptop" },
                "price": { "example": 1299.99 },
                "category": { "example": "Laptops" },
                "brand": { "example": "Dell" },
                "stock": { "example": 50 },
                "SKU": { "example": "GL-2024-001" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Products"],
        "description": "Retrieve a product by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Products"],
        "description": "Update a product by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "example": "Updated Product" },
                "price": { "example": 999.99 }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Products"],
        "description": "Delete a product by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "description": "Retrieve all categories",
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "OK" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Categories"],
        "description": "Create a new category",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "example": "Gaming" },
                "description": { "example": "Gaming devices and accessories" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "description": "Retrieve a category by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Categories"],
        "description": "Update a category by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "example": "Updated Category" },
                "description": { "example": "Updated description" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "description": "Delete a category by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  }
}